Now also stop the game if the application is killed. Use the latest version of messages. Protect arguments for the video server with quotes.

+++ src/orwell/game/Game.cpp
@@ -110,15 +110,17 @@ void Game::start()
 			char * aTempName = tmpnam(nullptr);
 			std::ofstream(aTempName).close();
 
-			aCommandLine << " cd server-web && make start ARGS='-u " << aRobot->getVideoUrl() << " -p " << aRobot->getVideoRetransmissionPort()
-					<< " --pid-file " << aTempName << "'";
-			ORWELL_LOG_INFO( "new tmp file : " << aTempName );
+			aCommandLine << " cd server-web && make start ARGS='-u \"" <<
+				aRobot->getVideoUrl() << "\" -p " <<
+				aRobot->getVideoRetransmissionPort() <<
+				" --pid-file " << aTempName << "'";
+			ORWELL_LOG_INFO("new tmp file : " << aTempName);
 			int aCode = system(aCommandLine.str().c_str());
-			ORWELL_LOG_INFO( "code at creation of webserver :" << aCode );
+			ORWELL_LOG_INFO("code at creation of webserver :" << aCode);
 
-			m_tmpFiles.push_back( aTempName );
+			m_tmpFiles.push_back(aTempName);
 		}
-		ORWELL_LOG_INFO( "game starts" );
+		ORWELL_LOG_INFO("game starts");
 		m_startTime = boost::posix_time::microsec_clock::local_time();
 		m_isRunning = true;
 	}
@@ -126,22 +128,25 @@ void Game::start()
 
 void Game::stop()
 {
-	for ( auto const aFileName: m_tmpFiles )
+	if (m_isRunning)
 	{
-		std::ifstream aFile(aFileName, std::ifstream::in);
-		int aPid = 0;
-		aFile >> aPid;
-		if (0 != aPid)
+		for ( auto const aFileName: m_tmpFiles )
 		{
-			kill(aPid, SIGABRT);
+			std::ifstream aFile(aFileName, std::ifstream::in);
+			int aPid = 0;
+			aFile >> aPid;
+			if (0 != aPid)
+			{
+				kill(aPid, SIGABRT);
+			}
+			else
+			{
+				ORWELL_LOG_ERROR("Could not kill a python web server.");
+			}
 		}
-		else
-		{
-			ORWELL_LOG_ERROR("Could not kill a python web server.");
-		}
+		ORWELL_LOG_INFO( "game stops" );
+		m_isRunning = false;
 	}
-	ORWELL_LOG_INFO( "game stops" );
-	m_isRunning = false;
 }
 
 bool Game::addRobot(
